convert to typescript with allow any
