+++ src/db.ts
@@ -1,7 +1,8 @@
 import mongoose from 'mongoose';
-import { Prompt } from './models/prompt.js';
+import { Prompt } from './models/prompt';
 import dotenv from 'dotenv';
-import { ResponseResult } from './models/response_result.js';
+import { ResponseResult } from './models/response_result';
+import assert from 'assert';
 dotenv.config();
 
 export const connectDb = async () => {
@@ -9,6 +10,7 @@ export const connectDb = async () => {
     console.log('connecting to db...');
     // suppress warnings
     mongoose.set('strictQuery', false);
+    assert(process.env.DB_URI, 'Missing DB URI');
     await mongoose.connect(process.env.DB_URI);
     console.log('connected to db');
   } catch (e) {
@@ -24,7 +26,7 @@ export const recordNewResponse = async ({ prompt, response, source, parentPrompt
 
 export const startNewPrompt = async ({ user, input }) => {
   const { _id } = await Prompt.create({ user, input });
-  return _id;
+  return _id.toString();
 }
 
 export const Rating = {
@@ -34,7 +36,7 @@ export const Rating = {
 
 export const incrementRatingCount = async ({responseId, rating}) => {
   const ratingKey = `correctAnswer.${rating}`;
-  const {correctAnswer} = await ResponseResult.findByIdAndUpdate(responseId, {$inc: {[ratingKey] : 1}}, {new: true}).lean().exec();
+  const {correctAnswer} = await ResponseResult.findByIdAndUpdate(responseId, {$inc: {[ratingKey] : 1}}, {new: true}).lean().exec() ?? {};
   return correctAnswer;
 }
 
@@ -45,7 +47,9 @@ export const updateSelectedAnswerSource = async ({promptId, source}) => {
 export const AnswerResult = {
   Answered: 'answered',
   RequestHelp: 'requestHelp'
-}
+} as const;
+
+export type AnswerResultChoice = typeof AnswerResult[keyof typeof AnswerResult]
 
 export const setPromptAnsweredResult = async ({promptId, answerResult}) => {
   const updateResult = answerResult === AnswerResult.Answered ? {answeredQuestion: true, requestedHelp: false} : {answeredQuestion: false, requestedHelp: true};
