+++ src/index.ts
@@ -1,6 +1,6 @@
 import express from 'express';
-import { startDiscord } from './bot.js';
-import { connectDb } from './db.js';
+import { startDiscord } from './bot';
+import { connectDb } from './db';
 
 import dotenv from 'dotenv';
 dotenv.config();
