+++ tsconfig.json
@@ -1,8 +1,25 @@
 {
-    "compilerOptions": {
-      "outDir": "./build",
-      "allowJs": true,
-      "target": "es5"
-    },
-    "include": ["./src/**/*"]
-  }
+  "compilerOptions": {
+    "allowJs": true,
+    "noImplicitAny": false,
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "lib": [
+      "es2022"
+    ],
+    "module": "commonjs",
+    "noUncheckedIndexedAccess": true,
+    "outDir": "./dist",
+    "skipLibCheck": true,
+    "strict": true,
+    "rootDir": "./src",
+    "target": "esnext",
+    "types": [
+      "node"
+    ]
+  },
+  "include": ["src/**/*"],
+  "exclude": [
+    "node_modules"
+  ]
+}
